@@ -4,23 +4,25 @@
   "private": true,
   "license": "MIT",
   "dependencies": {
-    "@types/jest": "24.0.11",
-    "@types/node": "11.11.6",
-    "@types/react": "16.8.8",
-    "@types/react-dom": "16.8.3",
-    "@xstate/react": "0.2.0",
-    "classnames": "^2.2.6",
+    "@types/jest": "26.0.23",
+    "@types/lodash": "^4.14.168",
+    "@types/node": "15.0.1",
+    "@types/react": "17.0.4",
+    "@types/react-dom": "17.0.3",
+    "@types/uuid": "^8.3.0",
+    "@xstate/react": "1.3.2",
+    "classnames": "^2.3.1",
     "emotion": "^10.0.9",
     "jsmidgen": "^0.1.5",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.21",
     "precision-inputs": "^0.2.5",
-    "react": "^16.8.5",
-    "react-aria-menubutton": "^6.1.0",
-    "react-dom": "^16.8.5",
-    "react-scripts": "2.1.8",
-    "typescript": "3.3.4000",
-    "uuid": "^3.3.2",
-    "xstate": "4.4.0"
+    "react": "^17.0.2",
+    "react-aria-menubutton": "7.0.1",
+    "react-dom": "^17.0.2",
+    "react-scripts": "4.0.3",
+    "typescript": "4.2.4",
+    "uuid": "^8.3.2",
+    "xstate": "4.18.0"
   },
   "scripts": {
     "start": "react-scripts start",
